reorganise directory for ParserTest
